+++ packages/merkleized-metadata/src/transform/index.ts
@@ -9,3 +9,5 @@ export {
   getCompactTypeTag,
   isVoidType,
 } from './typeUtils';
+export { decodeAndCollectLeaves, ExtrinsicDecodingError } from './decoder';
+export { lookupConstant } from './utils';
